App: Ensure EPTID is unique per RC service.
Previously we simply passed through the IdP generated EPTID value.
This however could have caused leakage of unique principals across
multiple services. Instead we now unpack the provided value
and recreate it as a unique value to satisfy the original intent of
EPTID.

We log the current EPTID against the generated version for future
retrieval incase of any security issue.