App: Further enhance EPTID generation to make us the source.
After further discussion we believed it to be more correct to have
the IdP section of EPTID set to the iss claim as we are modifying
the opaque part of the string. We further modifed the string to make
the SP section map to the aud claim.

This will ensure services wanting to determine who issued the EPTID
in the future will speak to AAF support rather then back to the
original IdP who would have no record of the EPTID value
being created.