Fix intermittent failure in export API tests
Because Sinatra::Base.set(..) operates at the class level, the state was
not reset between test runs. For certain random seeds, this caused the
export API to be disabled prior to tests which required it to be
enabled. Since those tests didn't turn the flag back on, the tests
failed.