Status: add new /status URL for Loadbalancer check
Add a /status URL that returns either 200 or 404 useful for a check by a
load-balancer.

If settings.status_disabled_file is set and the file exists, then this
instance is marked down and returns 404 at the /status URL (but still
handles all other requests normally).  This marks it as off-line for the
load balancer performing the check.

The status_disabled_file name defaults to:
    '/opt/rapidconnect/application/run/disabled'