Ensure user sessions contain all expected claims
Rapid Connect documentation lists the following attributes as being in
within the claim `https://aaf.edu.au/attributes` and describes them as
'SHOULD' exist.

1. cn
2. mail
3. displayname
4. edupersontargetedid
5. edupersonscopedaffiliation

To assist downstream developers this change catches sessions where
these attributes do not exist instead of passing them through to
connected applications.

Users who fall into this situation, it is not expected to occur often,
are provided with a request to contact AAF support and some debug data
to assist. AAF Support will then need to correct the issue as a teir 3
support activity with the backend IdP staff.