Implements support for Freshdesk SSO
The actual Freshdesk side has a number of issues here that are less
than ideal including:

* There appears to be no way for them to prevent replay attacks within
the 30 minute window the hash generated here is valid for
* This data being sent as name+val params isn't scalable.
* Why is the secret part of the message???????

Details at:
https://support.freshdesk.com/support/solutions/articles/31166-single-sign-on-remote-authentication-in-freshdesk

If this wasn't strictly over HTTPS I wouldn't be allowing this to go
into production for the AAF, as it stands it is barely passable.

Recommend that AAF sends the vendor a feature request to properly
support sending this data within a JWT like what out previous support
tool, Zendesk did.

Assuming the above request will be implemented this code is considered
as a temporary solution until such time as we can resume
simply sending JWT.