Mail: disable opportunistic STARTTLS (localhost)
Both RapidConnect and the Mail library default to opportunistic STARTTLS - use if supported.

On CentOS, Postfix does not support STARTLS so this is ignored.

On Ubuntu with REANNZ baseline applied, exim4 supports STARTTLS, but uses a
self-signed cert issued in the name hosts FQDN - which then does not match
"localhost" (and would not be trusted anyway because of being self-signed).

While this is not ideal situation, work around this by disabling STARTTLS - the connections
are hard-coded to localhost anyway.