AttributesClaim: allow ignoring eptid Redis set errors
When running in Master-Slave mode, the Slave will talk to a read-only Redis instance.

When a new user accesses a service for the first time (so targeted ID for the service
does not exist in Redis yet), the targetedID would be computed and stored in Redis.

But on the Slave instance, the storing would fail.

Handle this edge case by allowing to ignore the SET error and continue, just logging a warning.

Add a new configuration key "ignore_eptid_set_errors" - this is to be set to true only on Slave hosts.