Don't use emails when generating targeted ids
Shibboleth provides eduPersonTargetedIDs (EPTIDs) to services so
they can track users across sessions. For more information, see
https://support.aaf.edu.au/support/solutions/articles/19000052509-edupersontargetedid.

EPTIDs have two properties relevant to this commit:
1. Privacy. EPTIDS must be unique per service. A user accessing two
   services has two EPTIDs: One for each service.
2. Persistence. Users must get the same EPTID each time they access
   a given service. Changes to other attributes, like email
   addresses, must not affect their EPTID.

To preserve user privacy, RapidConnect must retarget EPTIDs.
Otherwise, it'd give the same one to every service provider.
Currently, RapidConnect uses the user's email when retargeting.
Unfortunately, this means RapidConnect must store the retargeted
EPTID. Otherwise, we'd lose EPTID persistence.

In this commit, we improve the situation. We remove email addresses
from the hashing logic. This makes the retargeting logic a pure
function of user and service.

To preserve persistence, we still need to look up EPTIDs
retargeted with the old logic. But because we now always get the
same EPTID from the same user and service, we don't need to store
more going forward. So we can remove that too.

This makes RapidConnect easier to migrate/upgrade, as it has one
less piece of live state to replicate. It also means we don't need
to worry about Redis running out of space.