fix: make sure sessions do expire
In 3f3f82e, a complex setup with custom classs RapidConnectMemcacheSession was
introduced to set expiry only for backend sessions but not for cookies - with
the premise that cookies without an expiry are session cookies that get
discarded when browser is closed.

This is stopped working when switching from memcache-client to dalli - the
overridden set_cookie method is not called from the dalli client.

Also, it no longer helps to set cookies as session - they are preserved by
browsers across restarts anyway.

So simplify this by removing the RapidConnectMemcacheSession and reverting to
simple :expire_after param given to Rack::Session::Dalli