new: logout: avoid an open redirect
Only accept a return URL passed to /logout if it is either:
* an HTTPS URL pointing to the same host
* a local path relative URL

Validate the URL through Ruby URI.parse as part of the process.

Log rejected URLs for audit purposes.

Note: keeping the return URL parameter for logout is useful for supporting
a logout hook for Shibboleth SP - and the local URL restriction will work with it well.

https://wiki.shibboleth.net/confluence/display/SP3/Notify